getting closer. undo for paste copy and cut are still broken with objects
getting closer.
undo for paste copy and cut are still broken with objects

2004-09-28  Jody Goldberg <jody@gnome.org>

	* src/commands.c (cmd_object_delete) : Take an optional descriptor
	  string.  All callers changed.

	* src/application.c (gnm_app_clipboard_cut_copy_obj) : Use
	  cmd_object_delete for cuts to get undo.

	* src/sheet-object.c (sheet_objects_relocate) : properly clear the
	  sheet before we unref an object.
	(sheet_objects_get) : use range_contained.
	(sheet_objects_clear) : ditto.  This simplifies undo/redo.

+++ src/gnumeric-pane.c
@@ -808,7 +808,7 @@ cb_control_point_event (FooCanvasItem *ctrl_pt, GdkEvent *event, GnmPane *pane)
 
 	case GDK_BUTTON_RELEASE:
 		if (pane->drag_object != so ||
-		    pane->drag_button != event->button.button)
+		    pane->drag_button != (int)event->button.button)
 			return FALSE;
 
 		cmd_object_move (WORKBOOK_CONTROL (scg_get_wbcg (scg)),
